pyproject.toml
[project]
name = "reporting-back"
version = "0.1.0"
description = "FastAPI backend for Excel analysis with AI"
authors = [{name = "Miguel Santana", email = "miguel@example.com"}]
requires-python = ">=3.12"
readme = "README.md"

dependencies = [
    # Web Framework
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "python-multipart==0.0.9",
    # Database
    "sqlalchemy==2.0.35",
    "alembic==1.13.2",
    "psycopg2-binary==2.9.9",
    "asyncpg==0.29.0",
    # Authentication & Security
    "pyjwt==2.8.0",
    "bcrypt==4.2.0",
    "email-validator==2.1.1",
    # Supabase Integration
    "supabase==2.7.4",
    # Background Tasks
    "celery[redis]==5.4.0",
    "redis==5.0.8",
    # Excel Processing
    "pandas==2.2.2",
    "openpyxl==3.1.5",
    # AI/LLM
    "langchain==0.3.0",
    "langgraph==0.2.20",
    "langchain-openai==0.2.0",
    # Configuration & Utilities
    "python-dotenv==1.0.1",
    "pydantic==2.9.0",
    "pydantic-settings==2.5.2",
    "httpx==0.27.2",
    "structlog==24.4.0",
    # Rate Limiting
    "slowapi==0.1.9",
    "langchain-anthropic>=0.3.15",
    "anthropic>=0.57.1",
    # Crypto Trading
    "alpaca-py>=0.28.0",
    "mplfinance>=0.12.10b0",
    "matplotlib>=3.8.0",
]

[tool.uv]
dev-dependencies = [
    # Testing
    "pytest==8.3.2",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "httpx==0.27.2",
    "factory-boy==3.3.1",
    "faker==28.4.1",
    # Code Quality
    "black==24.8.0",
    "ruff==0.6.5",
    "mypy==1.11.2",
    # Development Tools
    "ipython==8.26.0",
    "rich==13.7.1",
    "pandas-stubs>=2.3.0.250703",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
check = "scripts.build_scripts:check"
format = "scripts.build_scripts:format"
test = "scripts.build_scripts:test"
build = "scripts.build_scripts:build"

[tool.ruff]
line-length = 180
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "B008",  # Do not perform function calls in argument defaults - FastAPI Query/Depends pattern
]

[tool.black]
line-length = 180
target-version = ['py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

# SQLAlchemy uses dynamic attribute access that mypy can't understand
# This is a known limitation and the recommended approach is to disable
# these specific error codes for SQLAlchemy models
[[tool.mypy.overrides]]
module = "src.modules.*.models"
disable_error_code = ["assignment", "arg-type", "var-annotated"]

[[tool.mypy.overrides]]
module = "src.modules.*.service"  
disable_error_code = ["assignment"]

[[tool.mypy.overrides]]
module = "src.modules.*.controller"
disable_error_code = ["arg-type"]
